requirements.txt
admin-totals==1.0.1
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
Django==3.1.7
django-background-tasks==1.2.5
django-compat==1.0.15
django-environ==0.4.5
djangorestframework==3.12.2
idna==2.10
kiwisolver==1.3.1
lighthouse==1.0.2
lxml==4.6.2
matplotlib==3.3.4
numpy==1.20.1
pandas==1.2.3
Pillow==8.1.2
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tqdm==4.59.0
urllib3==1.26.3
xlrd==2.0.1
XlsxWriter==1.3.7
